feat: complete log data and download file

+++ include/TESTER_LOGGING.h
@@ -16,7 +16,7 @@ void appendFile(fs::FS &fs, const char * path, const char * message);
 /*-------------------function that self added-------------------*/
 
 void newFileInit();
-void logData(uint64_t cycleTtime);
+void logData(uint8_t cycleTtime);
 void endLogging();
 void downLogFile();
